Split tests into own files
